+++ src/main/java/com/ezycollect/payment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
 package com.ezycollect.payment.repository;
 
-import com.ezycollect.payment.model.Payment;
+import com.ezycollect.payment.dto.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
