+++ src/main/java/com/ezycollect/payment/repository/WebhookRepository.java
@@ -1,6 +1,6 @@
 package com.ezycollect.payment.repository;
 
-import com.ezycollect.payment.model.Webhook;
+import com.ezycollect.payment.dto.model.Webhook;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
