+++ src/main/java/com/ezycollect/payment/service/PaymentServiceImpl.java
@@ -3,8 +3,8 @@ package com.ezycollect.payment.service;
 import com.ezycollect.payment.dto.CreatePaymentRequest;
 import com.ezycollect.payment.exception.DatabaseException;
 import com.ezycollect.payment.exception.RetriesExhaustedException;
-import com.ezycollect.payment.mapper.PaymentMapper;
-import com.ezycollect.payment.model.Payment;
+import com.ezycollect.payment.dto.mapper.PaymentMapper;
+import com.ezycollect.payment.dto.model.Payment;
 import com.ezycollect.payment.repository.PaymentRepository;
 import com.ezycollect.payment.util.EncryptionUtil;
 import lombok.extern.slf4j.Slf4j;
