+++ src/main/java/com/ezycollect/payment/service/WebhookNotificationService.java
@@ -1,7 +1,7 @@
 package com.ezycollect.payment.service;
 
-import com.ezycollect.payment.model.Payment;
-import com.ezycollect.payment.model.Webhook;
+import com.ezycollect.payment.dto.model.Payment;
+import com.ezycollect.payment.dto.model.Webhook;
 
 import java.util.concurrent.CompletableFuture;
