+++ src/main/java/com/ezycollect/payment/service/WebhookService.java
@@ -1,6 +1,6 @@
 package com.ezycollect.payment.service;
 
-import com.ezycollect.payment.model.Payment;
+import com.ezycollect.payment.dto.model.Payment;
 import org.springframework.http.ResponseEntity;
 
 public interface WebhookService {
